+++ src/pages/nft-collections/[collectionAddress]/index.tsx
@@ -1,0 +1,3 @@
+import NftCollectionDetails from "lib/pages/nft-collection-details";
+
+export default NftCollectionDetails;