+++ src/pages/nft-collections/index.tsx
@@ -1,0 +1,3 @@
+import NftCollections from "lib/pages/nft-collections";
+
+export default NftCollections;